add reports nominas

+++ src/app/protected/home/components/company/components/nominas/nominas.component.ts
@@ -1,4 +1,4 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, computed, effect, OnInit, signal } from '@angular/core';
 import { Empleado } from '../personal/models/empleado.model';
 import { ActivatedRoute, Router } from '@angular/router';
 import { NominasService } from './nominas.service';
@@ -7,7 +7,10 @@ import { forkJoin } from 'rxjs';
 import { Arl, CalculoNomina, Nomina, NominaResumen } from './models/nomina.model';
 import { DialogService, DynamicDialogRef } from 'primeng/dynamicdialog';
 import { FormNominaComponent } from './form-nomina/form-nomina.component';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
+import { ReportePrimaComponent } from './reporte-prima/reporte-prima.component';
+import { ReporteCesantiasComponent } from './reporte-cesantias/reporte-cesantias.component';
+import { ReporteVacacionesComponent } from './reporte-vacaciones/reporte-vacaciones.component';
 
 @Component({
   selector: 'app-nominas',
@@ -30,6 +33,91 @@ export class NominasComponent implements OnInit {
 
   onViewAllCompany: boolean = false;
 
+  informeOptions: any[] | undefined;
+
+  selectInforme = signal<string>('');
+
+  informeEffect = effect(() => {
+    if(this.selectInforme() === 'reporte de primas'){
+      this.refFormUserAcces = this.dialogService.open(ReportePrimaComponent, {
+        header: 'Generar reporte Primna',
+        contentStyle: { overflow: 'auto' },
+        data: {
+          idEmpleado: this.selectedEmpleado?.id,
+        },
+        width: '300px'
+    });
+      this.refFormUserAcces.onClose.subscribe((data: any) => {
+          const periodo = data.periodo === 'Primer semestre' ? 1 : 2;
+          const anio = data.anio.getFullYear();
+          if (data) {
+            this.nominasService.getResumenPrimas(this.selectedEmpleado.id,anio, periodo).subscribe(
+              {
+                next: (resumenPrimas: any) => {
+                  this.generarInformePrimas(resumenPrimas, this.selectedEmpleado.nombresApellidos, this.selectedEmpleado.nombresApellidos);
+                },
+                error: (error) => {
+                  this.messageService.add({severity:'error', summary: 'Error', detail: error.error});
+                }
+              }
+            );
+        }
+      });
+    }
+    if(this.selectInforme() === 'reporte de cesantias'){
+      this.refFormUserAcces = this.dialogService.open(ReporteCesantiasComponent, {
+        header: 'Generar reporte cesantias',
+        contentStyle: { overflow: 'auto' },
+        data: {
+          idEmpleado: this.selectedEmpleado?.id,
+        },
+        width: '300px'
+    });
+      this.refFormUserAcces.onClose.subscribe((data: any) => {
+          const anio = data.anio.getFullYear();
+          if (data) {
+            this.nominasService.getResimenCesantias(this.selectedEmpleado.id,anio).subscribe(
+              {
+                next: (resumenPrimas: any) => {
+                  this.generarInformeCesantias(resumenPrimas, this.selectedEmpleado.nombresApellidos, this.selectedEmpleado.nombresApellidos);
+                },
+                error: (error) => {
+                  this.messageService.add({severity:'error', summary: 'Error', detail: error.error});
+                }
+              }
+            );
+        }
+      });
+    }
+    if(this.selectInforme() === 'reporte de vacaciones'){
+      this.refFormUserAcces = this.dialogService.open(ReporteVacacionesComponent, {
+        header: 'Generar reporte vacaciones',
+        contentStyle: { overflow: 'auto' },
+        data: {
+          idEmpleado: this.selectedEmpleado?.id,
+        },
+        width: '300px'
+    });
+      this.refFormUserAcces.onClose.subscribe((data: any) => {
+          //formato de fechas de Date a 2025-01-01
+          const inicio = data.date1.toISOString().split('T')[0];
+          const fin = data.date2.toISOString().split('T')[0];
+          if (data) {
+            this.nominasService.getResumenVacaciones(this.selectedEmpleado.id,inicio,fin).subscribe(
+              {
+                next: (resumenPrimas: any) => {
+                  this.generarInformeVacaciones(resumenPrimas, this.selectedEmpleado.nombresApellidos, this.selectedEmpleado.nombresApellidos);
+                },
+                error: (error) => {
+                  this.messageService.add({severity:'error', summary: 'Error', detail: error.error});
+                }
+              }
+            );
+        }
+      });
+    }
+  });
+
   constructor(
     private router: Router,
     private activatedRouter: ActivatedRoute,
@@ -37,9 +125,21 @@ export class NominasComponent implements OnInit {
     private empleadoService: EmpleadoService,
     private dialogService: DialogService,
     private messageService: MessageService,
+    private confirmationService: ConfirmationService,
   ) { }
 
   ngOnInit(): void {
+    this.informeOptions = [
+      {
+          cname: 'reporte de primas',
+      },
+      {
+          cname: 'reporte de cesantias',
+      },
+      {
+          cname: 'reporte de vacaciones',
+      }
+  ];
         forkJoin([
           this.empleadoService.getEmpleados(),
           this.nominasService.getARL()
@@ -159,17 +259,61 @@ export class NominasComponent implements OnInit {
   }
 
   eliminarNomina(nomina: NominaResumen){
-    this.nominasService.deleteNomina(nomina.id).subscribe({
-      next: () => {
-        this.messageService.add({severity:'success', summary: 'Nomina eliminada', detail: 'La nomina ha sido eliminada correctamente'});
-        this.nominasService.getNominasByIdEmpleado(this.selectedEmpleado.id).subscribe((nominas: NominaResumen[]) => {
-          this.nominas = nominas;
+    this.confirmationService.confirm({
+      message: `Desea eliminar esta nomina?`,
+      header: 'Eliminar nomina',
+      icon: 'pi pi-info-circle',
+      acceptButtonStyleClass:"p-button-danger p-button-text",
+      rejectButtonStyleClass:"p-button-text p-button-text",
+      acceptIcon:"none",
+      rejectIcon:"none",
+      acceptLabel:"Eliminar",
+      rejectLabel:"Cancelar",
+      accept: () => {
+        this.nominasService.deleteNomina(nomina.id).subscribe({
+          next: () => {
+            this.messageService.add({severity:'success', summary: 'Nomina eliminada', detail: 'La nomina ha sido eliminada correctamente'});
+            this.nominasService.getNominasByIdEmpleado(this.selectedEmpleado.id).subscribe((nominas: NominaResumen[]) => {
+              this.nominas = nominas;
+            });
+          },
+          error: (error) => {
+            this.messageService.add({severity:'error', summary: 'Error', detail: error.error});
+          }
         });
-      },
-      error: (error) => {
-        this.messageService.add({severity:'error', summary: 'Error', detail: error.error});
       }
     });
+  }
+
+  firmarNomina(nomina: NominaResumen){
+    this.confirmationService.confirm({
+      message: `una vez confirmada la nomina no podra editar ni eliminar el registro`,
+      header: 'Confirmar nomna',
+      icon: 'pi pi-info-circle',
+      acceptButtonStyleClass:"p-button-success p-button-text",
+      rejectButtonStyleClass:"p-button-text p-button-text",
+      acceptIcon:"none",
+      rejectIcon:"none",
+      acceptLabel:"Confirmar",
+      rejectLabel:"Cancelar",
+      accept: () => {
+        this.nominasService.firmarNomina(nomina.id).subscribe({
+          next: () => {
+            this.messageService.add({severity:'success', summary: 'Nomina firmada', detail: 'La nomina ha sido confirmada correctamente'});
+            this.nominasService.getNominasByIdEmpleado(this.selectedEmpleado.id).subscribe((nominas: NominaResumen[]) => {
+              this.nominas = nominas;
+            });
+          },
+          error: (error) => {
+            this.messageService.add({severity:'error', summary: 'Error', detail: error.error});
+          }
+        });
+      }
+    });
+  }
+
+  generarInforme(event){
+    this.selectInforme.set(event?.value?.cname);
   }
 
   setOnViewAllCompanyTrue(){
@@ -208,6 +352,18 @@ export class NominasComponent implements OnInit {
       const date = new Date(year, month - 1, day);
       const fechaAjustada = nominaResumen.fraccionMes ? `${year}-${month}/${day > 15 ? 2 : 1}` : `${year}-${month}`;
       this.nominasService.generatePdfNomina(nomina, 'nomina.pdf', this.selectedEmpleado.nombresApellidos + (this.selectedEmpleado.tipoNroDocumento ? ' ' + this.selectedEmpleado.tipoNroDocumento : ''), fechaAjustada);
+    }
+
+    generarInformePrimas(resumenPrimas: any, filename: string, empleado: string,){
+      this.nominasService.generatePdfReportePrimas(resumenPrimas, filename, empleado);
+    }
+
+    generarInformeCesantias(resumenCesantias: any, filename: string, empleado: string){
+      this.nominasService.generatePdfReporteCesantias(resumenCesantias, filename, empleado);
+    }
+
+    generarInformeVacaciones(resumenVacaciones: any, filename: string, empleado: string){
+      this.nominasService.generatePdfReporteVacaciones(resumenVacaciones, filename, empleado);
     }
 
 }
